start of cleaner version
