+++ charybdis/server/package.json
@@ -13,6 +13,9 @@
     "@inquirer/prompts": "^7.3.1",
     "@inquirer/type": "^3.0.4",
     "@types/figlet": "^1.7.0",
+    "cfonts": "^3.3.0",
+    "chalk": "^5.4.1",
+    "enquirer": "^2.4.1",
     "figlet": "^1.8.0",
     "gradient-string": "^3.0.0"
   },
